requirements.txt
Babel==2.3.4
Django==1.9.12
Django-Select2==5.8.10
Pillow==3.4.2
Unidecode==0.04.19
YURL==0.13
aldryn-apphooks-config==0.2.7
aldryn-boilerplates==0.7.4
aldryn-bootstrap3==1.1.2
aldryn-categories==1.0.3
aldryn-common==1.0.4
aldryn-newsblog==1.3.0
aldryn-people==1.2.2
aldryn-reversion==1.0.9
aldryn-translation-tools==0.2.1
backport-collections==0.1
cmsplugin-filer==1.1.3
django-appconf==1.0.2
django-appdata==0.1.6
django-bootstrap3==6.2.2
django-classy-tags==0.8.0
django-cms==3.4.1
django-durationfield==0.5.2
django-filer==1.2.5
django-formtools==1.0
django-jquery==3.1.0
django-mptt==0.8.6
django-parler==1.7
django-phonenumber-field==1.1.0
django-polymorphic==0.8.1
django-reversion==1.10.2
django-sekizai==0.10.0
django-sortedm2m==1.3.2
django-taggit==0.21.3
django-treebeard==4.1.0
djangocms-admin-style==1.2.6.2
djangocms-attributes-field==0.1.2
djangocms-column==1.7.0
djangocms-googlemap==1.0.1
djangocms-link==2.0.3
djangocms-snippet==1.9.2
djangocms-style==2.0.1
djangocms-text-ckeditor==3.3.1
djangocms-video==2.0.3
easy-thumbnails==2.3
html5lib==0.9999999
lxml==3.6.4
phonenumbers==7.7.5
phonenumberslite==7.7.5
psycopg2==2.6.2
python-dateutil==2.6.0
python-slugify==1.2.0
pytz==2016.7
six==1.10.0
